add sbig cooler (and report bug...)

+++ control/drivers/sx/SxCamera.cpp
@@ -175,7 +175,7 @@ SxCamera::~SxCamera() {
  * \param ccdindex	Index of the CCD to open.
  * \return A smart pointer to a CCD
  */
-CcdPtr	SxCamera::getCcd(int ccdindex) {
+CcdPtr	SxCamera::getCcd(size_t ccdindex) {
 	if (ccdindex >= nCcds()) {
 		debug(LOG_ERR, DEBUG_LOG, 0, "ccd id %d out of range",
 			ccdindex);
